chore(headless): rename navigator context provider file (#5191)
https://coveord.atlassian.net/browse/KIT-4138

So that I can sleep tonight.
