+++ packages/headless/src/case-assist.index.ts
@@ -35,7 +35,7 @@ export type {
 export type {LoggerOptions} from './app/logger.js';
 
 export type {LogLevel} from './app/logger.js';
-export type {NavigatorContext} from './app/navigatorContextProvider.js';
+export type {NavigatorContext} from './app/navigator-context-provider.js';
 
 // Case Assist Action Loaders
 export * from './features/case-input/case-input-actions-loader.js';
