+++ packages/headless/src/features/commerce/facets/facet-search-set/regular/commerce-regular-facet-search-request-builder.ts
@@ -1,5 +1,5 @@
 import {CommerceFacetSearchRequest} from '../../../../../api/commerce/facet-search/facet-search-request.js';
-import {NavigatorContext} from '../../../../../app/navigatorContextProvider.js';
+import {NavigatorContext} from '../../../../../app/navigator-context-provider.js';
 import {buildFilterableCommerceAPIRequest} from '../../../common/filterable-commerce-api-request-builder.js';
 import {getFacetIdWithoutCommerceFieldSuggestionNamespace} from '../commerce-facet-search-actions.js';
 import {StateNeededForRegularFacetSearch} from './commerce-regular-facet-search-state.js';
