+++ packages/headless/src/features/commerce/search/search-actions-thunk-processor.test.ts
@@ -4,7 +4,7 @@ import {Mock} from 'vitest';
 import {CommerceAPIClient} from '../../../api/commerce/commerce-api-client.js';
 import {CommerceAPIErrorStatusResponse} from '../../../api/commerce/commerce-api-error-response.js';
 import {SearchCommerceSuccessResponse} from '../../../api/commerce/search/response.js';
-import {defaultNodeJSNavigatorContextProvider} from '../../../app/navigatorContextProvider.js';
+import {defaultNodeJSNavigatorContextProvider} from '../../../app/navigator-context-provider.js';
 import {CommerceAppState} from '../../../state/commerce-app-state.js';
 import {buildSearchResponse} from '../../../test/mock-commerce-search.js';
 import {buildMockCommerceState} from '../../../test/mock-commerce-state.js';
