+++ packages/headless/src/features/facets/facet-search-set/category/category-facet-search-request-builder.ts
@@ -1,6 +1,6 @@
 import {getSearchApiBaseUrl} from '../../../../api/platform-client.js';
 import {CategoryFacetSearchRequest} from '../../../../api/search/facet-search/category-facet-search/category-facet-search-request.js';
-import {NavigatorContext} from '../../../../app/navigatorContextProvider.js';
+import {NavigatorContext} from '../../../../app/navigator-context-provider.js';
 import {buildSearchRequest} from '../../../search/search-request.js';
 import {CategoryFacetRequest} from '../../category-facet-set/interfaces/request.js';
 import {StateNeededForCategoryFacetSearch} from '../generic/generic-facet-search-state.js';
