+++ packages/headless/src/features/search-and-folding/search-and-folding-request.ts
@@ -2,7 +2,7 @@ import {isNullOrUndefined} from '@coveo/bueno';
 import {EventDescription} from 'coveo.analytics';
 import {getSearchApiBaseUrl} from '../../api/platform-client.js';
 import {SearchRequest} from '../../api/search/search/search-request.js';
-import {NavigatorContext} from '../../app/navigatorContextProvider.js';
+import {NavigatorContext} from '../../app/navigator-context-provider.js';
 import {SearchAppState} from '../../state/search-app-state.js';
 import {ConfigurationSection} from '../../state/state-sections.js';
 import {fromAnalyticsStateToAnalyticsParams} from '../configuration/analytics-params.js';
