+++ packages/headless/src/features/standalone-search-box-set/standalone-search-box-set-actions.ts
@@ -7,7 +7,7 @@ import {
   AsyncThunkSearchOptions,
   isErrorResponse,
 } from '../../api/search/search-api-client.js';
-import {NavigatorContext} from '../../app/navigatorContextProvider.js';
+import {NavigatorContext} from '../../app/navigator-context-provider.js';
 import {
   ConfigurationSection,
   ContextSection,
